requirements.txt
typer[all]>=0.12.3
httpx>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
platformdirs>=4.2.2
#openai>=1.40.0
jinja2>=3.1.4
python-multipart>=0.0.9
ollama>=0.6.1
